Bubble_sort_command_line
Write a program that takes nine integer as its arguments from command line. The first eight arguments are going to be sorted using the bubble sort. The ninth argument is going to be used as the keyword that will be searched for by your program using binary search. Your program should display the original unsorted inputs (note: without the 9th argument, which is the searching keyword), then displays the sorted inputs, and then displays the position of the keyword in the sorted array.